Create simulate_data.R
Builds a sample data set with 2 variables that have 6 months of dependence, but one is only viewed quarterly. Then builds a quantile regression for a variable of choice using the (unobserved) monthly series.

Estimates quantile reg using the monthly series and the mixed frequency series.

Currently not bootstrapped, but can be expended to include bootstrap tests.